[swi_glu/swi_glu_grad] kernel ut修复；host ut修复；增加example



# message auto-generated for no-merge-commit merge:
!53 merge master into master

[swi_glu/swi_glu_grad] kernel ut修复；host ut修复；增加example

Created-by: merlinsantosh
Commit-by: merlinsantosh
Merged-by: turing_project1
Description: ## 描述
[swi_glu/swi_glu_grad] kernel ut修复；host ut修复；增加example

## 关联的Issue
N/A

## 测试
bash build.sh --pkg --soc=ascend910b --vendor_name=custom --ops=swi_glu
bash build.sh --run_example swi_glu eager cust
bash build.sh --opkernel -u --ops=swi_glu
bash build.sh --ophost -u --ops=swi_glu

bash build.sh --pkg --soc=ascend910b --vendor_name=custom --ops=swi_glu_grad
bash build.sh --run_example swi_glu_grad eager cust
bash build.sh --opkernel -u --ops=swi_glu_grad
bash build.sh --ophost -u --ops=swi_glu_grad

## 文档更新
更新了算子README.md文件

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [x] 文档更新
- [ ] 其他，请描述：补充ut




See merge request: cann/ops-nn!53

+++ activation/swi_glu/op_host/swi_glu_tiling.cpp
@@ -470,7 +470,7 @@ ge::graphStatus Tiling4SwiGlu(gert::TilingContext* context) {
     }
     uint64_t ubSize = 0;
     platformInfo.GetCoreMemSize(platform_ascendc::CoreMemType::UB, ubSize);
-    OP_TILING_CHECK((totalCore < 0 || ubSize <= UB_RESERVED_BUFF),
+    OP_TILING_CHECK((static_cast<int>(totalCore) < 0 || ubSize <= UB_RESERVED_BUFF),
         OP_LOGE(context->GetNodeName(), "Compile Info is invalid, coreNum:%u, ubSize:%lu", totalCore, ubSize),
         return ge::GRAPH_PARAM_INVALID);
     ubSize -= UB_RESERVED_BUFF;
